Update on "Merge Tensor and Variable."
All Tensors are Variables now, they just happen to have requires_grad=False. Tensors ALWAYS have `VariableTensorId` in their type set.

When constructing this patch, I had to make decisions about what I would fix in this patch, and what I would leave for follow up PRs. Here is the cleanup that happens in this patch:

- The `is_variable` property is removed from TensorOptions. I removed this immediately because unlike Tensor::is_variable, TensorOptions::is_variable doesn't respect our VariableTensorId thread-local state. This means that there were a bunch of places where TensorOptions::is_variable was false, which is obviously bogus in the world when tensor and variable are merged. Instead of keeping the method as a function that always returns true, I just opted to remove it entirely (it's not public API.) All places we set `is_variable` are deleted.
  - Knock on effect: there is no longer a separate DeprecatedTypeProperties for the variable and non-variable versions of type.
  - Knock on effect: instead of asserting on TensorOptions::is_variable, instead we just test `at::impl::variable_is_excluded()`
- There is now only one copy of the cuDNN RNN dropout cache, not two (I'm not sure why we had two to begin with)

Some cleanup that doesn't happen in this patch:
- Eliminating unnecessary uses of `make_variable`
- Eliminating `Tensor::is_variable`

The most subtle part of this patch is retaining tracing behavior: the fact that everything is a Variable means that more code gets routed to VariableType than before; this can change traces. I identified two places where we didn't appropriately turn off VariableType, mostly factory functions:

- `torch.tensor` must turn off VariableType before invoking `at::empty` to construct the tensor, as it subsequently does direct data access
- `tensor_slow` (invoked when you pass a Python scalar to a tensor argument) must turn off VariableType before calling `scalar_to_tensor` so the scalar gets traced as constant, rather than as a call to `scalar_to_tensor`.

Honestly, these are all giant hacks, and should be replaced with a more specialized guard that just toggles tracing.


Differential Revision: [D18171156](https://our.internmc.facebook.com/intern/diff/D18171156)

[ghstack-poisoned]

+++ test/onnx/test_utility_funs.py
@@ -137,8 +137,23 @@ class TestUtilityFuns(TestCase):
     def test_constant_fold_concat(self):
         class ConcatModule(torch.nn.Module):
             def forward(self, x):
-                a = torch.tensor([[1., 2., 3.]])
-                b = torch.tensor([[4., 5., 6.]])
+                # Why did I insert a Cast here?  There appears to be intentional
+                # behavior in ONNX constant folding where constant tensors which
+                # are not attached to any known to be foldable onnx
+                # operations don't get extracted into the initializer graph.  So
+                # without these casts, we will actually fail to pull out one of
+                # the constants, thus failing constant folding.  I think the
+                # test is wrong but I don't have time to write a more correct
+                # test (I think the right way to go about the test is to setup
+                # a predicate for what invariant graphs should hold after
+                # constant folding, and then verify this predicate holds.
+                # I think the asserts below are an attempt at this predicate,
+                # but it is not right!)
+                #
+                # More commentary at
+                # https://github.com/pytorch/pytorch/pull/18698/files#r340107552
+                a = torch.tensor([[1., 2., 3.]]).to(torch.float)
+                b = torch.tensor([[4., 5., 6.]]).to(torch.float)
                 c = torch.cat((a, b), 0)
                 return b + c
